+++ src/main/java/mx/axxib/aforedigitalgt/dal/NotificacionPagosRepo.java
@@ -1,5 +1,6 @@
 package mx.axxib.aforedigitalgt.dal;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.ParameterMode;
@@ -34,8 +35,8 @@ public class NotificacionPagosRepo extends RepoBase {
 			StoredProcedureQuery query = entityManager.createStoredProcedureQuery(storedFullName, "LlenaInfo");
 
 			query.registerStoredProcedureParameter("P_FECHA_CONSULTA", Date.class, ParameterMode.IN);
-			query.registerStoredProcedureParameter("P_TOT_TITULOS", Integer.class, ParameterMode.OUT);
-			query.registerStoredProcedureParameter("P_TOT_PESOS", Integer.class, ParameterMode.OUT);
+			query.registerStoredProcedureParameter("P_TOT_TITULOS", BigDecimal.class, ParameterMode.OUT);
+			query.registerStoredProcedureParameter("P_TOT_PESOS", BigDecimal.class, ParameterMode.OUT);
 			query.registerStoredProcedureParameter("CP_DATOS", void.class, ParameterMode.REF_CURSOR);
 			query.registerStoredProcedureParameter("P_MENSAJE", String.class, ParameterMode.OUT);
 			query.registerStoredProcedureParameter("P_ESTATUS", Integer.class, ParameterMode.OUT);
@@ -46,8 +47,8 @@ public class NotificacionPagosRepo extends RepoBase {
 			Object cursor = query.getOutputParameterValue("CP_DATOS");
 			if (cursor != null) {
 				res.setDatos(query.getResultList());
-				res.setTotTitulos((Integer)query.getOutputParameterValue("P_TOT_TITULOS"));
-				res.setTotPesos((Integer)query.getOutputParameterValue("P_TOT_PESOS"));
+				res.setTotTitulos((BigDecimal)query.getOutputParameterValue("P_TOT_TITULOS"));
+				res.setTotPesos((BigDecimal)query.getOutputParameterValue("P_TOT_PESOS"));
 				res.setMensaje((String)query.getOutputParameterValue("P_MENSAJE"));
 				res.setEstatus((Integer)query.getOutputParameterValue("P_ESTATUS"));
 			}
